OpenPaaS/esn-frontend-admin#67 - Move imap-item-service To inbox.libs
