+++ src/esn.inbox.libs/app/services/config/config.spec.js
@@ -9,7 +9,7 @@ describe('The inboxConfig factory', function() {
   var $rootScope, inboxConfig, config;
 
   beforeEach(function() {
-    angular.mock.module('linagora.esn.unifiedinbox', function($provide) {
+    angular.mock.module('esn.inbox.libs', function($provide) {
       config = config || {};
 
       $provide.value('esnConfig', function(key, defaultValue) {
