+++ src/esn.inbox.libs/app/services/mailboxes/special-mailboxes.spec.js
@@ -8,7 +8,7 @@ describe('The inboxSpecialMailboxes factory', function() {
 
   var inboxSpecialMailboxes;
 
-  beforeEach(angular.mock.module('linagora.esn.unifiedinbox'));
+  beforeEach(angular.mock.module('esn.inbox.libs'));
 
   beforeEach(angular.mock.inject(function(_inboxSpecialMailboxes_) {
     inboxSpecialMailboxes = _inboxSpecialMailboxes_;
