+++ src/esn.inbox.libs/views/partials/quotes/forwardText.pug
