+++ src/linagora.esn.unifiedinbox/app/components/sidebar/folder-settings/folder-settings.controller.js
@@ -1,7 +1,5 @@
 'use strict';
 
-require('../../../services/jmap-item/jmap-item-service.js');
-
 angular.module('linagora.esn.unifiedinbox')
   .controller('inboxFolderSettingsController', inboxFolderSettingsController);
